Gather scatter pattern test and fixes (#57)
* Gather scatter pattern test and fixes

This commit builds on top of #46, with an update on tests and scripts. It also makes two minor changes to UCXVan:

* ZPull does not require a preceding ZPush
* pull request still pass the correct value of val_len in msg.meta
* added a thread to ensure the order of push/pull request messages upon reception
* It also adds a USE_CUDA=1 compilation option.

- fix w_pool_ for pull
- fix val len bug
- add USE_CUDA=1 compilation option
- unset export UCX_TLS=ib,cuda
- support 48 bit keys

+++ src/customer.cc
@@ -12,8 +12,8 @@
 #include <chrono>
 
 namespace ps {
-const int Node::kEmpty = std::numeric_limits<int>::max();
-const int Meta::kEmpty = std::numeric_limits<int>::max();
+const int Node::kEmpty = std::numeric_limits<short>::max();
+const int Meta::kEmpty = std::numeric_limits<short>::max();
 
 Customer::Customer(int app_id, int customer_id, const Customer::RecvHandle& recv_handle)
     : app_id_(app_id), customer_id_(customer_id), recv_handle_(recv_handle) {
